Greedy online alog now handles edge cases

+++ structs.go
@@ -3,10 +3,10 @@ package stripack
 
 
 
-func Array(dx, dy int) [][]int {
-    a := make([][]int, dx) 
+func Array(dx, dy int) [][]int64 {
+    a := make([][]int64, dx) 
     for i := range a {
-        a[i] = make([]int, dy) 
+        a[i] = make([]int64, dy) 
         for j := range a[i] {
             a[i][j] = -1
         }
